Add spec to show overridden vs replaced attr accessors
In the current implementation the AttributeMethod module is shared
between all members of a virtus enabled class inheritance chain.

So the virtus_define_{writer,reader}_method calls are replacing the attr
accessor implementations of the base class. The define_method calls of
the descendants do have the same object as the receiver.
This was not spotted since redefinition of an attribute with the same
name is unusual.

The next commit is a fix, I isolated this one to support easy
verification of my claims.