Make descendants include their attributes accessor modules.
By including their attributes accessor module on the first call to `ClassMethods#attribute`, I suspect their is a small possibility of weird behavior around attribute accessor method precedence, but I don't think it will be an issue in real usage. 

If it accessor method precedence is a real issue with this approach, an alternative would be to have the descendant include its attribute accessor methods module in a new `ClassMethods#inherited` method.