Fix spec failure due to stub conversion to string.
This test was failing because the name attribute was specified as a
string and the subject.name was being compared to the name stub (of
class RSpec::Mock) rather than another string and thus was failing.

It appears that this has come up in the past (specifically #78) and is
caused by the fact that RSpec::Mock implements `to_str`.