Fix but in coerce when Rails 5 params are passed in
Since parameters in Rails 5 no longer sub class from Hash and blow up
when you try to call to_hash on them when they are not permitted the
coerce method in Virtus was no longer working. Since the hash here is
not mass assigned but instead only permits known attributes it is safe
to call `.to_unsafe_hash` if it is defined on the attributes. If not we
fall back to the previous behavior.

This also adds specs to capture the bahavior I was seeing on Rails 5.